split `bounds` module into `const_param_ty` and `copy` modules
